requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
pillow
vercel_blob
